% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aceCodeCompletion}
\alias{aceCodeCompletion}
\title{Enable Code Completion for an Ace Code Input}
\usage{
aceCodeCompletion(inputId, fromList = NULL,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{The id of the input object}

\item{fromList}{The list of code completion words that will be categorized as \code{names}.
This is intended to provide a way to offer completion for functions that use NSE (non-standard
evaluation). This can be a static character vector/\code{list} or a \code{reactive} that returns
such.}

\item{session}{The \code{session} object passed to function given to shinyServer}
}
\value{
An observer reference class object that is responsible for offering code completion.
See \code{\link[shiny]{observe}} for more details.
}
\description{
Enable Code Completion for an Ace Code Input
}

