% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aceCodeInput}
\alias{aceCodeInput}
\title{Create an Ace code input control}
\usage{
aceCodeInput(inputId, label, code = NULL, hint = c("autocomplete",
  "anyword"))
}
\arguments{
\item{inputId}{Input variable to assign the control's value to}

\item{label}{Display label for the control}

\item{code}{Initial code text}

\item{hint}{Types of code compeltion to enable}
}
\value{
An Ace code input control that can be added to a UI definition
}
\description{
Create an input control for entry of R code text values. This is very similar to a
\code{\link[shiny]{textInput}} but has additional IDE look and feel plus code
completion that comes with Ace code editor.
}

