% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updateAceCodeInput}
\alias{updateAceCodeInput}
\title{Change the value of an Ace code input on the client}
\usage{
updateAceCodeInput(session, inputId, label = NULL, code = NULL,
  hint = NULL)
}
\arguments{
\item{session}{The \code{session} object passed to function given to ShinyServer}

\item{inputId}{The id of the input object}

\item{label}{The label to set for the input object}

\item{code}{The code text to set for the input object}

\item{hint}{The type of code completion to enable for the input object}
}
\description{
Change the value of an Ace code input on the client
}

